Added database models
