+++ README.md
@@ -101,3 +101,8 @@ $ flask --app webapp/app run --debug
   }
 }
 ```
+## Notes
+Background tasks are not enabled in debug mode by default, as they cause migrations to hang. They can be enabled by setting `ENABLE_TASKS`.
+```bash
+export ENABLE_TASKS=True
+```
