+++ requirements.txt
@@ -9,6 +9,7 @@ Jinja2==3.1.4
 MarkupSafe==2.1.5
 multiprocess==0.70.16
 pydantic==2.8.2
+psycopg2-binary==2.9.9
 PyYAML==6.0.1
 Flask-Migrate==4.0.7
 Flask-SQLAlchemy==3.1.1
