+++ webapp/__init__.py
@@ -2,7 +2,7 @@ from flask import Flask
 
 from webapp.cache import init_cache
 from webapp.context import base_context
-from webapp.database import init_db
+from webapp.models import init_db
 from webapp.sso import init_sso
 from webapp.tasks import init_tasks
 
@@ -25,4 +25,5 @@ def create_app():
 
     # Initialize tasks
     init_tasks(app)
+
     return app
