--- webapp/database.py
@@ -1,16 +1,0 @@
-from flask_migrate import Migrate
-from flask_sqlalchemy import SQLAlchemy
-from sqlalchemy.orm import DeclarativeBase
-
-
-class Base(DeclarativeBase):
-    pass
-
-
-db = SQLAlchemy(model_class=Base)
-
-
-def init_db(app):
-    Migrate(app, db)
-    app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///project.db"
-    db.init_app(app)
